# deploy/office-assistant/hermes/office-gateway.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hermes Office Gateway",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://office-gateway:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/host": {
      "get": {
        "operationId": "office_host",
        "summary": "Lab VM Docker processes plus CPU, memory, and disk"
      }
    },
    "/v1/services/{name}": {
      "get": {
        "operationId": "office_service",
        "summary": "Single service health",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/actions/propose": {
      "post": {
        "operationId": "office_propose_action",
        "summary": "Propose an allowlisted write (does not execute)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "target"],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "restart_service",
                      "scale_replicas",
                      "clear_queue",
                      "set_feature_flag"
                    ]
                  },
                  "target": {
                    "type": "string"
                  },
                  "params": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/execute": {
      "post": {
        "operationId": "office_execute_action",
        "summary": "Execute a previously proposed action_id after user APPROVE",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action_id"],
                "properties": {
                  "action_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}